chore: support sdk v3
