+++ src/dynamodb/index.ts
@@ -1,0 +1,4 @@
+import * as dynamodbv2 from "./v2";
+import * as dynamodbv3 from "./v3";
+
+export { dynamodbv2, dynamodbv3 };
