+++ tests/configs/tables-function-async/jest-dynalite-config.js
@@ -8,5 +8,5 @@ module.exports = {
     await sleep(300);
     return tables;
   },
-  basePort: 8000,
+  basePort: 10500,
 };
