+++ tests/configs/tables-function/jest-dynalite-config.js
@@ -2,5 +2,5 @@ const tables = require("../tables");
 
 module.exports = {
   tables: () => tables,
-  basePort: 8000,
+  basePort: 10500,
 };
