+++ tests/jest-dynalite-config.js
@@ -4,5 +4,5 @@ const tables = require("./configs/tables");
 // More advanced test configs can be found in .jest/configs
 module.exports = {
   tables,
-  basePort: 8000,
+  basePort: 10500,
 };
