+++ tests/setups/setupDynamodbV2.ts
@@ -1,0 +1,5 @@
+jest.resetModules();
+jest.mock("../../src/utils", () => ({
+  ...(jest.requireActual("../../src/utils") as any),
+  hasV3: () => false,
+}));
